10주차 > basic > SeachDialog 값 일부 메모이제이션, 인풋 필드들 컴포넌트 분리 및 메모이제이션
