+++ packages/assignment-10/src/basic/ScheduleTables.tsx
@@ -2,15 +2,12 @@ import { Button, ButtonGroup, Flex, Heading, Stack } from "@chakra-ui/react";
 import { useState } from "react";
 import { useScheduleContext } from "./ScheduleContext.tsx";
 import ScheduleTable from "./ScheduleTable.tsx";
-import SearchDialog from "./SearchDialog.tsx";
+import SearchDialog from "./SearchDialog/index.tsx";
+import { SearchInfo } from "./types.ts";
 
 export const ScheduleTables = () => {
   const { schedulesMap, setSchedulesMap } = useScheduleContext();
-  const [searchInfo, setSearchInfo] = useState<{
-    tableId: string;
-    day?: string;
-    time?: number;
-  } | null>(null);
+  const [searchInfo, setSearchInfo] = useState<SearchInfo | null>(null);
 
   const disabledRemoveButton = Object.keys(schedulesMap).length === 1;
