+++ packages/assignment-10/src/basic/SearchDialog/Inputs/DayCheckBoxGroup.tsx
@@ -1,0 +1,39 @@
+import {
+  Checkbox,
+  CheckboxGroup,
+  FormControl,
+  FormLabel,
+  HStack,
+} from "@chakra-ui/react";
+import { memo } from "react";
+import { DAY_LABELS } from "../../constants";
+import { SearchOption } from "../../types";
+
+interface Props {
+  changeSearchOption: (
+    field: keyof SearchOption,
+    value: SearchOption[typeof field]
+  ) => void;
+  value: string[];
+}
+const DayCheckBoxGroup = ({ value, changeSearchOption }: Props) => {
+  return (
+    <FormControl>
+      <FormLabel>요일</FormLabel>
+      <CheckboxGroup
+        value={value}
+        onChange={(value) => changeSearchOption("days", value as string[])}
+      >
+        <HStack spacing={4}>
+          {DAY_LABELS.map((day) => (
+            <Checkbox key={day} value={day}>
+              {day}
+            </Checkbox>
+          ))}
+        </HStack>
+      </CheckboxGroup>
+    </FormControl>
+  );
+};
+
+export default memo(DayCheckBoxGroup);
