+++ packages/assignment-10/src/basic/types.ts
@@ -13,3 +13,18 @@ export interface Schedule {
   range: number[];
   room?: string;
 }
+
+export interface SearchInfo {
+  tableId: string;
+  day?: string;
+  time?: number;
+}
+
+export interface SearchOption {
+  query?: string;
+  grades: number[];
+  days: string[];
+  times: number[];
+  majors: string[];
+  credits?: number;
+}
